requirements.txt
requests
annotated-types==0.7.0
asgiref==3.9.1
cffi==2.0.0
contextlib2==21.6.0
cryptography==45.0.7
Django==5.2.6
django-cors-headers==4.8.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
injector==0.22.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
